+++ catalog/src/main/java/it/eng/catalog/model/DataService.java
@@ -12,18 +12,15 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedBy;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.annotation.Version;
-import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonProperty.Access;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 
@@ -74,10 +71,6 @@ public class DataService implements Serializable {
     @JsonProperty(DSpaceConstants.DCAT_ENDPOINT_URL)
     private String endpointURL;
     
-    @JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class) //, property = "id"
-    @JsonProperty(DSpaceConstants.DCAT_SERVES_DATASET)
-    @DBRef
-    private Set<Dataset> servesDataset;
     @JsonIgnore
     @CreatedBy
     private String createdBy;
@@ -175,12 +168,6 @@ public class DataService implements Serializable {
             return this;
         }
 
-        @JsonProperty(DSpaceConstants.DCAT_SERVES_DATASET)
-        public Builder servesDataset(Set<Dataset> servesDataset) {
-            service.servesDataset = servesDataset;
-            return this;
-        }
-
         @JsonProperty("createdBy")
         public Builder createdBy(String createdBy) {
             service.createdBy = createdBy;
@@ -242,7 +229,6 @@ public class DataService implements Serializable {
          .title(updatedDataService.getTitle() != null ? updatedDataService.getTitle() : this.title)
          .endpointDescription(updatedDataService.getEndpointDescription() != null ? updatedDataService.getEndpointDescription() : this.endpointDescription)
          .endpointURL(updatedDataService.getEndpointURL() != null ? updatedDataService.getEndpointURL() : this.endpointURL)
-         .servesDataset(updatedDataService.getServesDataset() != null ? updatedDataService.getServesDataset() : this.servesDataset)
          .build();
   }
 }